adding checks for pre-set env variale
